requirements.txt
absl-py>=0.12.0
apache_beam
earthengine-api
geopandas
google_apitools
google-cloud-aiplatform
google-cloud-bigquery-storage
numpy
opencv-python
pandas
pillow
pyproj
pytest
rasterio
requests
rio-cogeo
shapely
sklearn
tensorflow
tensorflow_probability
tqdm
tensorflow-addons
